Renamed TextFileReader to TestFileReader #1103
- moved from integartion test package to
  test package
- as mentioned in review of PR #1105
- added test framework as test dependency for
  commons model
- TestFileReader has static methods
- TextFileReader has no static methods
  (so mockable...)
