+++ sechub-commons-model/build.gradle
@@ -12,6 +12,7 @@ dependencies{
     
     api project(':sechub-commons-core')
     
+    testImplementation project(':sechub-testframework')
     testImplementation spring_boot_dependency.junit_jupiter
     
     testImplementation spring_boot_dependency.mockito_core
