+++ sechub-commons-model/src/test/java/com/mercedesbenz/sechub/commons/model/SecHubScanConfigurationTest.java
@@ -7,7 +7,7 @@ import java.io.File;
 
 import org.junit.jupiter.api.Test;
 
-import com.mercedesbenz.sechub.commons.TextFileReader;
+import com.mercedesbenz.sechub.test.TestFileReader;
 
 class SecHubScanConfigurationTest {
 
@@ -15,8 +15,7 @@ class SecHubScanConfigurationTest {
     void sechub_job_config_example1_JSON_can_be_desrialized_and_contains_expected_login_url() {
 
         /* prepare */
-        TextFileReader reader = new TextFileReader();
-        String json = reader.loadTextFile(new File("./src/test/resources/sechub_config_example1.json"));
+        String json = TestFileReader.loadTextFile(new File("./src/test/resources/sechub_config_example1.json"));
 
         /* execute */
         SecHubScanConfiguration scanConfig = SecHubScanConfiguration.createFromJSON(json);
